Сделаны все процессы для домашки 10.2
