+++ README.md
@@ -11,3 +11,7 @@ EXECUTED
 2. Принимает на вход список словарей и возвращает новый список, в котором исходные словари отсортированы по убыванию даты (ключ 
 date
 ). Функция принимает два аргумента, второй необязательный задает порядок сортировки (убывание, возрастание)
+
+## Тесты:
+
+Реализованы тесты с покрытием более 80%
