+++ pyproject.toml
@@ -18,6 +18,7 @@ isort = "^5.13.2"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.2.1"
+pytest-cov = "^5.0.0"
 
 [build-system]
 requires = ["poetry-core"]
